+++ internal/auth-service/repository/find.go
@@ -13,7 +13,7 @@ func (r *authRepo) FindUserByEmail(ctx context.Context, email string) (*model.Us
 	return &user, nil
 }
 
-func (r *authRepo) FindUserByID(ctx context.Context, id int64) (*model.User, error) {
+func (r *authRepo) FindUserByID(ctx context.Context, id int) (*model.User, error) {
 	var user model.User
 
 	if err := r.db.Where("id = ?", id).First(&user).Error; err != nil {
