+++ internal/common/sql_model.go
@@ -5,7 +5,7 @@ import (
 )
 
 type SqlModel struct {
-	ID        int64     `json:"id" gorm:"column:id"`
+	ID        int       `json:"id" gorm:"column:id"`
 	UpdatedAt time.Time `gorm:"column:updated_at" json:"updated_at,omitempty"`
 	CreatedAt time.Time `gorm:"column:created_at" json:"created_at,omitempty"`
 }
